refactor: 패키지 리팩토링 shape 과 관련된 height width 이동
