+++ src/main/java/ladder/domain/ladder/shape/LadderShapeInfo.java
@@ -1,7 +1,5 @@
 package ladder.domain.ladder.shape;
 
-import ladder.domain.ladder.Height;
-import ladder.domain.ladder.PillarCount;
 import ladder.domain.player.Players;
 import ladder.domain.prize.Prizes;
 import ladder.exception.ErrorMessage;
