+++ src/test/java/ladder/domain/LadderGameTest.java
@@ -1,7 +1,7 @@
 package ladder.domain;
 
 import ladder.domain.dto.LadderMatchResult;
-import ladder.domain.ladder.Height;
+import ladder.domain.ladder.shape.Height;
 import ladder.domain.ladder.shape.LadderShapeInfo;
 import ladder.domain.player.Player;
 import ladder.domain.player.Players;
