+++ src/test/java/ladder/domain/ladder/LadderTest.java
@@ -1,5 +1,6 @@
 package ladder.domain.ladder;
 
+import ladder.domain.ladder.shape.Height;
 import ladder.domain.ladder.shape.LadderShapeInfo;
 import ladder.domain.player.Players;
 import ladder.domain.prize.Prizes;
