+++ src/test/java/ladder/domain/ladder/shape/PillarCountTest.java
@@ -1,5 +1,6 @@
-package ladder.domain.ladder;
+package ladder.domain.ladder.shape;
 
+import ladder.domain.ladder.shape.PillarCount;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
